Tests/LibWeb/Text/expected/focus-shouldIndicateFocus.txt

INPUT: rgb(0, 0, 0) auto 3px
BUTTON: rgb(0, 0, 0) 0px
